revert to nixpkgs
plenary has forced errors during lua module import checks, which are
performed during the build process for the drv. this is a nix-related
error, because it tests plugins by importing everything.

I don't know how these checks are being enforced when pulling the plugin
from git instead of a package.

https://nixos.org/manual/nixpkgs/unstable/#testing-neovim-plugins-neovim-require-check