disable broken plugins
mini-animate did not have issues as a git plugin `pkgs.neovimPlugins`
but now it is broken.

mini-align was also working that way i think, but i forget if i ever
enabled it.

dial just has some error with the default augend group thing. this might
be unrelated to plugin loading.